Rename build_avoiding_target -> build_repaired_targets
The tool repairs ANY rule violation (road-border crossing, moving/static
collision), not just avoidance; the old name was misleading. Rename the module to
match its public build_repaired_targets() entrypoint and update the workflow
runner, tests, and README. build_avoiding_target.py stays as a thin deprecation
shim (re-exports main) so an in-flight workflow that already spawned
python -m ...build_avoiding_target keeps resolving; it will be removed after.

Smoke: python -m rlvr.autoresearch.tools.build_repaired_targets on 16 mined
credit windows (K=8, grpo_temperature) produced 5 repaired scenes with correct
per-scene ego_shape + target_gt_disagreement; full R2LPL unit suite green.
